etl/pyproject.toml
[project]
name = "fpl-etl"
version = "0.1.0"
description = "Fantasy Premier League ETL Pipeline"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "requests>=2.31.0",
    "snowflake-connector-python>=3.7.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "pyarrow>=15.0.0",
]

[project.scripts]
fpl-etl = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []